+++ DirectX 11 VS2022/Graphics/Camera.cpp
@@ -118,14 +118,14 @@ void Camera::SetLookAtPos(XMFLOAT3 lookAtPos)
 	float pitch = 0.0f;
 	if (lookAtPos.y != 0.0f)
 	{
-		const float distance = sqrt(lookAtPos.x * lookAtPos.x + lookAtPos.z * lookAtPos.z);
-		pitch = atan(lookAtPos.y / distance);
+		const float distance = (float)sqrt(lookAtPos.x * lookAtPos.x + lookAtPos.z * lookAtPos.z);
+		pitch = (float)atan(lookAtPos.y / distance);
 	}
 
 	float yaw = 0.0f;
 	if (lookAtPos.x != 0.0f)
 	{
-		yaw = atan(lookAtPos.x / lookAtPos.z);
+		yaw = (float)atan(lookAtPos.x / lookAtPos.z);
 	}
 	if (lookAtPos.z > 0)
 		yaw += XM_PI;
